Unit Test - Assert Types

+++ Demo/Calculator.cs
@@ -7,7 +7,7 @@
             return v1 + v2;
         }
 
-        public double Divide(double v1, double v2)
+        public decimal Divide(decimal v1, decimal v2)
         {
             return v1 / v2;
         }
